fix(ci): python-coverage-shards — split oversized worker/cleanup.py under the 1500-line guardrail
The issue #399 deferred Claude auth-overlay umount work grew
src/awf/control/worker/cleanup.py to 1696 lines, tripping the maintainability
guardrail test_first_party_code_files_stay_under_line_limit (MAX 1500) and
failing the coverage shard.

Extract the cohesive auth-overlay umount subsystem (the bounded inline retry
_teardown_terminal_auth_overlay, the deferred re-sweep functions, the marker
record/count/guard helpers, and all _TERMINAL_AUTH_OVERLAY_UNMOUNT_* constants)
into a new sibling module awf.control.worker.cleanup_auth_overlay with behavior
unchanged. cleanup.py (990 lines) re-imports the three symbols it still uses
(_teardown_terminal_auth_overlay and the PENDING / RETRY_SCAN_FAILED constants);
the new module is 747 lines. Repoint the mixin delegations and the dedicated
auth-overlay tests (which monkeypatch module globals, so they must target the
defining module) to the new module.
